ENT-1454_Moving_clients_endpoints
