+++ lib/bwapi/client/admin/clients.rb
@@ -1,4 +1,3 @@
-require 'bwapi/client/admin/clients/sub_clients'
 require 'bwapi/client/admin/clients/users'
 require 'bwapi/client/admin/clients/user_groups'
 
@@ -21,7 +20,6 @@ module BWAPI
           get "admin/clients/#{client_id}/usage/daily", opts
         end
 
-        include BWAPI::Client::Admin::Clients::SubClients
         include BWAPI::Client::Admin::Clients::Users
         include BWAPI::Client::Admin::Clients::UserGroups
       end
